+++ cryptofeed/callback.py
@@ -107,9 +107,21 @@ class BalancesCallback(Callback):
     pass
 
 
-class AccTransactionsCallback(Callback):
+class TransactionsCallback(Callback):
     pass
 
 
 class UserFillsCallback(Callback):
+    pass
+
+
+class L1BookCallback(Callback):
+    pass
+
+
+class UserDataCallback(Callback):
+    pass
+
+
+class LastPriceCallback(Callback):
     pass