+++ cryptofeed/exceptions.py
@@ -52,3 +52,7 @@ class ConnectionClosed(Exception):
 
 class UnexpectedMessage(Exception):
     pass
+
+
+class ConnectionExists(Exception):
+    pass